pyproject.toml
[tool.poetry]
name = "test_task"
version = "0.1.0"
description = ""
authors = ["MankoMax <fanatscape@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.11.1"
python-dotenv = "^0.21.0"
requests = "^2.28.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
